# packages/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: RelayFlow API
  version: 0.1.0
  description: API contract for the RelayFlow backend.
servers:
  - url: http://localhost:8080
paths:
...
  /profile:
    get:
      operationId: getProfile
      summary: Get the current user's profile and security settings
      tags: [Profile]
      responses:
        "200":
          description: Current profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProfileResponse"
    patch:
      operationId: updateProfile
      summary: Update profile details and preferences
      tags: [Profile]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileRequest"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProfileResponse"
    delete:
      operationId: deleteAccount
      summary: Delete the current user's account and invalidate the session
      tags: [Profile]
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeleteAccountRequest"
      responses:
        "204":
          description: Account deleted
...
  /profile/2fa/enable:
    post:
      operationId: enable2FA
      summary: Enable TOTP two-factor authentication after OTP verification
      tags: [Profile]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OtpRequest"
      responses:
        "204":
          description: Two-factor authentication enabled
  /profile/2fa/disable:
    post:
      operationId: disable2FA
      summary: Disable TOTP two-factor authentication after OTP verification
      tags: [Profile]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OtpRequest"
      responses:
        "204":
          description: Two-factor authentication disabled

  # ── Workspaces ────────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-Api-Key
...
  schemas:
...
    UpdateProfileRequest:
      type: object
      required: [displayName, receiveEmailUpdates]
      properties:
        displayName:
          type: string
          maxLength: 200
        receiveEmailUpdates:
          type: boolean
...
    DeleteAccountRequest:
      type: object
      properties:
        password:
          type: string
          nullable: true
          description: Required for email/password accounts; OAuth users may omit it.
    OtpRequest:
      type: object
      required: [otp]
      properties:
        otp:
          type: string
...
    ProfileResponse:
      type: object
      required: [userId, email, receiveEmailUpdates, twoFactorEnabled, providers]
      properties:
        userId:
          type: string
          format: uuid
        email:
          type: string
          format: email
        displayName:
          type: string
          nullable: true
        avatarUrl:
          type: string
          nullable: true
        receiveEmailUpdates:
          type: boolean
        twoFactorEnabled:
          type: boolean
        providers:
          type: array
          items:
            type: string
            enum: [GOOGLE, EMAIL]
...